Remove cudautils.py
