+++ python/cudf/cudf/core/udf/utils.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 
 import functools
 import os
+from pickle import dumps
 from typing import TYPE_CHECKING, Any
 
 import cachetools
@@ -19,16 +20,15 @@ from numba.types import CPointer, Record, Tuple
 import rmm
 
 from cudf._lib import strings_udf
-from cudf.core.column.column import ColumnBase, as_column
-from cudf.core.dtypes import dtype
+from cudf.core.udf.masked_typing import MaskedType
 from cudf.core.udf.strings_typing import (
     string_view,
     udf_string,
 )
-from cudf.utils import cudautils
 from cudf.utils._numba import _get_ptx_file
 from cudf.utils.dtypes import (
     BOOL_TYPES,
+    CUDF_STRING_DTYPE,
     DATETIME_TYPES,
     NUMERIC_TYPES,
     STRING_TYPES,
@@ -46,8 +46,6 @@ if TYPE_CHECKING:
 # Maximum size of a string column is 2 GiB
 _STRINGS_UDF_DEFAULT_HEAP_SIZE = os.environ.get("STRINGS_UDF_HEAP_SIZE", 2**31)
 _heap_size = 0
-_cudf_str_dtype = dtype(str)
-
 
 JIT_SUPPORTED_TYPES = (
     NUMERIC_TYPES
@@ -61,6 +59,12 @@ MASK_BITSIZE = np.dtype("int32").itemsize * 8
 
 precompiled: cachetools.LRUCache = cachetools.LRUCache(maxsize=32)
 launch_arg_getters: dict[Any, Any] = {}
+
+# This cache is keyed on the (signature, code, closure variables) of UDFs, so
+# it can hit for distinct functions that are similar. The lru_cache wrapping
+# compile_udf misses for these similar functions, but doesn't need to serialize
+# closure variables to check for a hit.
+_udf_code_cache: cachetools.LRUCache = cachetools.LRUCache(maxsize=32)
 
 
 @functools.cache
@@ -104,7 +108,7 @@ def _masked_array_type_from_col(col):
     array of bools representing a mask.
     """
 
-    if col.dtype == _cudf_str_dtype:
+    if col.dtype == CUDF_STRING_DTYPE:
         col_type = CPointer(string_view)
     else:
         nb_scalar_ty = numpy_support.from_dtype(col.dtype)
@@ -142,6 +146,76 @@ def _mask_get(mask, pos):
     return (mask[pos // MASK_BITSIZE] >> (pos % MASK_BITSIZE)) & 1
 
 
+def make_cache_key(udf, sig):
+    """
+    Build a cache key for a user defined function. Used to avoid
+    recompiling the same function for the same set of types
+    """
+    codebytes = udf.__code__.co_code
+    constants = udf.__code__.co_consts
+    names = udf.__code__.co_names
+
+    if udf.__closure__ is not None:
+        cvars = tuple(x.cell_contents for x in udf.__closure__)
+        cvarbytes = dumps(cvars)
+    else:
+        cvarbytes = b""
+
+    return names, constants, codebytes, cvarbytes, sig
+
+
+def compile_udf(udf, type_signature):
+    """Compile ``udf`` with `numba`
+
+    Compile a python callable function ``udf`` with
+    `numba.cuda.compile_ptx_for_current_device(device=True)` using
+    ``type_signature`` into CUDA PTX together with the generated output type.
+
+    The output is expected to be passed to the PTX parser in `libcudf`
+    to generate a CUDA device function to be inlined into CUDA kernels,
+    compiled at runtime and launched.
+
+    Parameters
+    ----------
+    udf:
+      a python callable function
+
+    type_signature:
+      a tuple that specifies types of each of the input parameters of ``udf``.
+      The types should be one in `numba.types` and could be converted from
+      numpy types with `numba.numpy_support.from_dtype(...)`.
+
+    Returns
+    -------
+    ptx_code:
+      The compiled CUDA PTX
+
+    output_type:
+      An numpy type
+
+    """
+    key = make_cache_key(udf, type_signature)
+    res = _udf_code_cache.get(key)
+    if res:
+        return res
+
+    # We haven't compiled a function like this before, so need to fall back to
+    # compilation with Numba
+    ptx_code, return_type = cuda.compile_ptx_for_current_device(
+        udf, type_signature, device=True
+    )
+    if not isinstance(return_type, MaskedType):
+        output_type = numpy_support.as_dtype(return_type).type
+    else:
+        output_type = return_type
+
+    # Populate the cache for this function
+    res = (ptx_code, output_type)
+    _udf_code_cache[key] = res
+
+    return res
+
+
 def _generate_cache_key(frame, func: Callable, args, suffix="__APPLY_UDF"):
     """Create a cache key that uniquely identifies a compilation.
 
@@ -152,9 +226,7 @@ def _generate_cache_key(frame, func: Callable, args, suffix="__APPLY_UDF"):
     """
     scalar_argtypes = tuple(typeof(arg) for arg in args)
     return (
-        *cudautils.make_cache_key(
-            func, tuple(_all_dtypes_from_frame(frame).values())
-        ),
+        make_cache_key(func, tuple(_all_dtypes_from_frame(frame).values())),
         *(col.mask is None for col in frame._columns),
         *frame._column_names,
         scalar_argtypes,
@@ -166,7 +238,7 @@ def _get_input_args_from_frame(fr: IndexedFrame) -> list:
     args: list[Buffer | tuple[Buffer, Buffer]] = []
     offsets = []
     for col in _supported_cols_from_frame(fr).values():
-        if col.dtype == _cudf_str_dtype:
+        if col.dtype == CUDF_STRING_DTYPE:
             data = column_to_string_view_array_init_heap(
                 col.to_pylibcudf(mode="read")
             )
@@ -184,17 +256,9 @@ def _get_input_args_from_frame(fr: IndexedFrame) -> list:
 
 
 def _return_arr_from_dtype(dtype, size):
-    if dtype == _cudf_str_dtype:
+    if dtype == CUDF_STRING_DTYPE:
         return rmm.DeviceBuffer(size=size * _get_extensionty_size(udf_string))
     return cp.empty(size, dtype=dtype)
-
-
-def _post_process_output_col(col, retty):
-    if retty == _cudf_str_dtype:
-        return ColumnBase.from_pylibcudf(
-            strings_udf.column_from_udf_string_array(col)
-        )
-    return as_column(col, retty)
 
 
 # The only supported data layout in NVVM.
