+++ python/cudf/cudf/tests/test_compile_udf.py
@@ -1,21 +1,21 @@
-# Copyright (c) 2021, NVIDIA CORPORATION.
+# Copyright (c) 2021-2025, NVIDIA CORPORATION.
 
 from numba import types
 
-from cudf.utils import cudautils
+from cudf.core.udf.utils import _udf_code_cache, compile_udf
 
 
 def setup_function():
-    cudautils._udf_code_cache.clear()
+    _udf_code_cache.clear()
 
 
 def assert_cache_size(size):
-    assert cudautils._udf_code_cache.currsize == size
+    assert _udf_code_cache.currsize == size
 
 
 def test_first_compile_sets_cache_entry():
     # The first compilation should put an entry in the cache
-    cudautils.compile_udf(lambda x: x + 1, (types.float32,))
+    compile_udf(lambda x: x + 1, (types.float32,))
     assert_cache_size(1)
 
 
@@ -23,10 +23,10 @@ def test_code_cache_same_code_different_function_hit():
     # Compilation of a distinct function with the same code and signature
     # should reuse the cached entry
 
-    cudautils.compile_udf(lambda x: x + 1, (types.float32,))
+    compile_udf(lambda x: x + 1, (types.float32,))
     assert_cache_size(1)
 
-    cudautils.compile_udf(lambda x: x + 1, (types.float32,))
+    compile_udf(lambda x: x + 1, (types.float32,))
     assert_cache_size(1)
 
 
@@ -34,10 +34,10 @@ def test_code_cache_different_types_miss():
     # Compilation of a distinct function with the same code but different types
     # should create an additional cache entry
 
-    cudautils.compile_udf(lambda x: x + 1, (types.float32,))
+    compile_udf(lambda x: x + 1, (types.float32,))
     assert_cache_size(1)
 
-    cudautils.compile_udf(lambda x: x + 1, (types.float64,))
+    compile_udf(lambda x: x + 1, (types.float64,))
     assert_cache_size(2)
 
 
@@ -49,10 +49,10 @@ def test_code_cache_different_cvars_miss():
     def gen_closure(y):
         return lambda x: x + y
 
-    cudautils.compile_udf(gen_closure(1), (types.float32,))
+    compile_udf(gen_closure(1), (types.float32,))
     assert_cache_size(1)
 
-    cudautils.compile_udf(gen_closure(2), (types.float32,))
+    compile_udf(gen_closure(2), (types.float32,))
     assert_cache_size(2)
 
 
@@ -62,6 +62,6 @@ def test_lambda_in_loop_code_cached():
     # ensuring that there is only one entry in the code cache after the loop.
 
     for i in range(3):
-        cudautils.compile_udf(lambda x: x + 1, (types.float32,))
+        compile_udf(lambda x: x + 1, (types.float32,))
 
     assert_cache_size(1)
